refactor: make plot views stateful
